requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.53.0
pydantic==2.5.0
pydantic-settings==2.1.0
reportlab==4.0.7
pypdf==5.1.0
python-docx==1.1.0
httpx==0.27.0
weasyprint==62.3
xhtml2pdf==0.2.16
azure-storage-blob==12.23.1
azure-identity==1.19.0

pytest==7.4.0
